+++ src/TNRAF/TNRAFDistanceQueryManagerWithMapping.cpp
@@ -15,6 +15,6 @@ TNRAFDistanceQueryManagerWithMapping::TNRAFDistanceQueryManagerWithMapping(Trans
 
 // Used to find the shortest distance from source to target where source and target are the original indices.
 //______________________________________________________________________________________________________________________
-double TNRAFDistanceQueryManagerWithMapping::findDistance(const long long unsigned int source, const long long unsigned int target) {
+unsigned int TNRAFDistanceQueryManagerWithMapping::findDistance(const long long unsigned int source, const long long unsigned int target) {
     return qm.findDistance(mapping.at(source), mapping.at(target));
 }
